Move GameState up to the forge.game module.
