+++ forge-gui-desktop/src/test/java/forge/net/HeadlessNetworkGuiGame.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 import forge.LobbyPlayer;
-import forge.ai.GameState;
+import forge.game.GameState;
 import forge.deck.CardPool;
 import forge.game.GameEntityView;
 import forge.game.card.CardView;
