+++ forge-gui/src/main/java/forge/gamemodes/net/server/RemoteClientGuiGame.java
@@ -1,7 +1,7 @@
 package forge.gamemodes.net.server;
 
 import forge.LobbyPlayer;
-import forge.ai.GameState;
+import forge.game.GameState;
 import forge.deck.CardPool;
 import forge.game.GameEntityView;
 import forge.game.event.GameEvent;
