+++ forge-gui/src/main/java/forge/gui/interfaces/IGuiGame.java
@@ -1,7 +1,7 @@
 package forge.gui.interfaces;
 
 import forge.LobbyPlayer;
-import forge.ai.GameState;
+import forge.game.GameState;
 import forge.deck.CardPool;
 import forge.game.GameEntityView;
 import forge.game.GameView;
